+++ .gitignore
@@ -3,4 +3,4 @@ gradle
 run
 build
 .idea
-/forge/.architectury-transformer/debug.log
+/neoforge/.architectury-transformer/debug.log
