--- forge/src/main/resources/META-INF/mods.toml
@@ -1,36 +1,0 @@
-modLoader = "javafml" #mandatory
-loaderVersion = "[43,)" #mandatory This is typically bumped every Minecraft version by Forge. See our download page for lists of versions.
-license = "${license}"
-issueTrackerURL = "${repository}/issues"
-[[mods]] #mandatory
-modId = "cobblemon_quests" #mandatory
-version = "${version}" #mandatory
-displayName = "${mod_name}" #mandatory
-updateJSONURL = "https://raw.githubusercontent.com/WinterWolfSV/Cobblemon_Quests/master/forge/src/main/resources/META-INF/forge_mod_versions.json" #optional
-displayURL = "https://modrinth.com/mod/cobblemon-quests/versions" #optional
-logoFile = "icon.png" #optional
-description = "${mod_description}"
-
-[[dependencies.cobblemon_quests]] #optional
-modId = "forge" #mandatory
-mandatory = true #mandatory
-versionRange = "[43,)" #mandatory
-ordering = "NONE"
-side = "BOTH"
-
-[[dependencies.cobblemon_quests]]
-modId = "minecraft"
-mandatory = true
-versionRange = "[1.20.1]"
-ordering = "NONE"
-side = "BOTH"
-
-[[dependencies.cobblemon_quests]]
-modId = "cobblemon"
-mandatory = true
-versionRange = "[1.4.0,)"
-
-[[dependencies.cobblemon_quests]]
-modId = "ftbquests"
-mandatory = true
-versionRange = "[2001.1.5,)"
