+++ neoforge/src/main/resources/icon.png
